+++ config/test.exs
@@ -6,7 +6,6 @@ config :td_dd, TdDdWeb.Endpoint,
   http: [port: 3005],
   server: true
 
-
 # Hashing algorithm just for testing porpouses
 config :td_dd, hashing_module: TrueBG.DummyHashing
 
@@ -15,7 +14,6 @@ config :logger, level: :warn
 
 # Configure your database
 config :td_dd, TdDd.Repo,
-  adapter: Ecto.Adapters.Postgres,
   username: "postgres",
   password: "postgres",
   database: "td_dd_test",
@@ -26,7 +24,8 @@ config :td_dd, :api_services_login,
   api_username: "api-admin",
   api_password: "apipass"
 
-config :td_dd, :auth_service, api_service: TdDdWeb.ApiServices.MockTdAuthService,
+config :td_dd, :auth_service,
+  api_service: TdDdWeb.ApiServices.MockTdAuthService,
   auth_host: "localhost",
   auth_port: "4001",
   auth_domain: ""
@@ -40,10 +39,12 @@ config :td_dd, :elasticsearch,
 config :td_dd, df_cache: TdPerms.MockDynamicFormCache
 config :td_dd, permission_resolver: TdDd.Permissions.MockPermissionResolver
 config :td_dd, taxonomy_cache: TdDd.MockTaxonomyCache
+config :td_dd, index_worker: TdDd.Search.MockIndexWorker
 
 config :td_perms, redis_host: "localhost"
 
-config :td_dd, :audit_service, api_service: TdDdWeb.ApiServices.MockTdAuditService,
+config :td_dd, :audit_service,
+  api_service: TdDdWeb.ApiServices.MockTdAuditService,
   audit_host: "localhost",
   audit_port: "4007",
   audit_domain: ""