+++ lib/td_dd_web/controllers/comment_controller.ex
@@ -19,7 +19,6 @@ defmodule TdDdWeb.CommentController do
   end
 
   swagger_path :index do
-    get "/comments"
     description "List Comments"
     response 200, "OK", Schema.ref(:CommentsResponse)
   end
@@ -29,7 +28,6 @@ defmodule TdDdWeb.CommentController do
   end
 
   swagger_path :create do
-    post "/comments"
     description "Creates Comments"
     produces "application/json"
     parameters do
@@ -47,13 +45,12 @@ defmodule TdDdWeb.CommentController do
       Audit.create_event(conn, audit, @events.create_comment)
       conn
       |> put_status(:created)
-      |> put_resp_header("location", comment_path(conn, :show, comment))
+      |> put_resp_header("location", Routes.comment_path(conn, :show, comment))
       |> render("show.json", comment: comment)
     end
   end
 
   swagger_path :show do
-    get "/comments/{id}"
     description "Show Comment"
     produces "application/json"
     parameters do
@@ -68,7 +65,6 @@ defmodule TdDdWeb.CommentController do
   end
 
   swagger_path :update do
-    patch "/comments/{id}"
     description "Update Comments"
     produces "application/json"
     parameters do
@@ -89,7 +85,6 @@ defmodule TdDdWeb.CommentController do
   end
 
   swagger_path :delete do
-    delete "/comments/{id}"
     description "Delete Comment"
     produces "application/json"
     parameters do
@@ -108,7 +103,6 @@ defmodule TdDdWeb.CommentController do
   end
 
   swagger_path :get_comment_data_field do
-    get "/data_fields/{data_field_id}/comment"
     description "Show Data Field Comment"
     produces "application/json"
     parameters do
@@ -124,7 +118,6 @@ defmodule TdDdWeb.CommentController do
   end
 
   swagger_path :get_comment_data_structure do
-    get "/data_structures/{data_structure_id}/comment"
     description "Show Data Structure Comment"
     produces "application/json"
     parameters do