+++ lib/td_dd_web/controllers/search_controller.ex
@@ -3,11 +3,11 @@ defmodule TdDdWeb.SearchController do
   import Canada, only: [can?: 2]
   use PhoenixSwagger
   alias TdDd.DataStructures.DataStructure
-  alias TdDd.Search.IndexWorker
   alias TdDdWeb.ErrorView
 
+  @index_worker Application.get_env(:td_dd, :index_worker)
+
   swagger_path :reindex_all do
-    get("/data_structures/search/reindex_all")
     description("Reindex all ES indexes with DB content")
     produces("application/json")
     response(202, "Accepted")
@@ -18,18 +18,20 @@ defmodule TdDdWeb.SearchController do
     user = conn.assigns[:current_user]
 
     with true <- can?(user, reindex_all(DataStructure)) do
-      IndexWorker.reindex()
+      @index_worker.reindex()
       send_resp(conn, :accepted, "")
     else
       false ->
         conn
         |> put_status(:forbidden)
-        |> render(ErrorView, "403.json")
+        |> put_view(ErrorView)
+        |> render("403.json")
 
       _error ->
         conn
         |> put_status(:internal_server_error)
-        |> render(ErrorView, "500.json")
+        |> put_view(ErrorView)
+        |> render("500.json")
     end
   end
 end