+++ test/td_dd_web/controllers/data_structure_filter_controller_test.exs
@@ -6,14 +6,13 @@ defmodule TdDdWeb.DataStructureFilterControllerTest do
   alias TdDd.Search.MockSearch
   alias TdDdWeb.ApiServices.MockTdAuditService
   alias TdDdWeb.ApiServices.MockTdAuthService
-
-  @df_cache Application.get_env(:td_dd, :df_cache)
+  alias TdPerms.MockDynamicFormCache
 
   setup_all do
-    start_supervised MockTdAuthService
-    start_supervised MockTdAuditService
-    start_supervised MockPermissionResolver
-    start_supervised(@df_cache)
+    start_supervised(MockTdAuthService)
+    start_supervised(MockTdAuditService)
+    start_supervised(MockPermissionResolver)
+    start_supervised(MockDynamicFormCache)
     :ok
   end
 
@@ -25,15 +24,14 @@ defmodule TdDdWeb.DataStructureFilterControllerTest do
   describe "index" do
     @tag :admin_authenticated
     test "lists all filters (admin user)", %{conn: conn} do
-      conn = get conn, data_structure_filter_path(conn, :index)
+      conn = get(conn, Routes.data_structure_filter_path(conn, :index))
       assert json_response(conn, 200)["data"] == MockSearch.get_filters(%{})
     end
 
     @tag authenticated_no_admin_user: @user_name
     test "lists all filters (non-admin user)", %{conn: conn} do
-      conn = get conn, data_structure_filter_path(conn, :index)
+      conn = get(conn, Routes.data_structure_filter_path(conn, :index))
       assert json_response(conn, 200)["data"] == %{}
     end
   end
-
 end